requirements.txt
google-genai
groq
requests
